remove global error handler
