+++ opentelemetry/src/global/internal_logging.rs
@@ -1,4 +1,14 @@
 #![allow(unused_macros)]
+
+#[cfg(feature = "logs")]
+use crate::logs::LogError;
+#[cfg(feature = "metrics")]
+use crate::metrics::MetricError;
+use crate::propagation::PropagationError;
+#[cfg(feature = "trace")]
+use crate::trace::TraceError;
+use std::sync::PoisonError;
+
 ///
 /// **Note**: These macros (`otel_info!`, `otel_warn!`, `otel_debug!`, and `otel_error!`) are intended to be used
 /// **internally within OpenTelemetry code** or for **custom exporters and processors**. They are not designed
@@ -155,4 +165,40 @@ macro_rules! otel_error {
             let _ = ($name, $($value),+); // Compiler will optimize this out as it's unused.
         }
     };
+}
+
+/// Wrapper for error from tracing, log, and metrics part of open telemetry.
+#[derive(thiserror::Error, Debug)]
+#[non_exhaustive]
+pub enum Error {
+    #[cfg(feature = "trace")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "trace")))]
+    #[error(transparent)]
+    /// Failed to export traces.
+    Trace(#[from] TraceError),
+    #[cfg(feature = "metrics")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "metrics")))]
+    #[error(transparent)]
+    /// An issue raised by the metrics module.
+    Metric(#[from] MetricError),
+
+    #[cfg(feature = "logs")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "logs")))]
+    #[error(transparent)]
+    /// Failed to export logs.
+    Log(#[from] LogError),
+
+    #[error(transparent)]
+    /// Error happens when injecting and extracting information using propagators.
+    Propagation(#[from] PropagationError),
+
+    #[error("{0}")]
+    /// Other types of failures not covered by the variants above.
+    Other(String),
+}
+
+impl<T> From<PoisonError<T>> for Error {
+    fn from(err: PoisonError<T>) -> Self {
+        Error::Other(err.to_string())
+    }
 }
