+++ opentelemetry/src/global/mod.rs
@@ -130,7 +130,6 @@
 //! [`MeterProvider`]: crate::metrics::MeterProvider
 //! [`set_meter_provider`]: crate::global::set_meter_provider
 
-mod error_handler;
 mod internal_logging;
 #[cfg(feature = "metrics")]
 mod metrics;
@@ -139,7 +138,7 @@ mod propagation;
 #[cfg(feature = "trace")]
 mod trace;
 
-pub use error_handler::{handle_error, set_error_handler, Error};
+pub use internal_logging::Error;
 #[cfg(feature = "metrics")]
 #[cfg_attr(docsrs, doc(cfg(feature = "metrics")))]
 pub use metrics::*;
